Crate subspace_sdk

Subspace SDK for easy running of both Subspace node and farmer

Modules

chain_spec

Subspace chain configurations.

farmer

Farmer related things located here

node

Node related things located here

Structs

ByteSize

Container for number of bytes.

FarmerBuilder

Module related to the farmer Builder for Config.

FarmerInfo

Module related to the farmer Farmer info

Multiaddr

Multiaddr is a wrapper around libp2p one

MultiaddrWithPeerId

Multiaddr with peer id

NodeBuilder

Builder for Config.

NodeInfo

Node state info

PlotDescription

Module related to the farmer Description of the plot

PublicKey

Public key type

Enums

Ss58ParsingError

An error type for SS58 decoding.

Type Definitions

Farmer

Subspace farmer type

Node

Subspace primary node

Plot

Subspace farmer’s plot