Module subspace_sdk::node

Node related things located here

Re-exports

pub use super::Node;

Modules

chain_spec

Subspace chain configurations.

domains

Module for executor and its domains

Structs

BaseBuilder

Builder for Base.

Constraints

Pruning constraints. If none are specified pruning is

Dsn

Node DSN builder

DsnBuilder

Builder for Dsn.

DsnOptions

Options for DSN

DsnShared

Shared Dsn structure between node and farmer

FarmerPieceCache

Piece cache with limited size where pieces closer to provided peer ID are retained.

ImplName

Type wrapper with default value for implementation name

ImplVersion

Type wrapper with default value for implementation version

InConnections

Wrapper with default value for number of incoming connections

ListenAddresses

Wrapper with default value for listen address

MaxSubsPerConn

Wrapper with default value for max subscriptions per connection

Network

Node network builder

NetworkBuilder

Builder for Network.

NodePieceGetter

Node piece getter (combines DSN and Farmer getters)

OffchainWorker

Offchain worker config

OffchainWorkerBuilder

Builder for OffchainWorker.

OutConnections

Wrapper with default value for number of outgoing connections

PendingInConnections

Wrapper with default value for number of pending incoming connections

PendingOutConnections

Wrapper with default value for number of pending outgoing connections

Rpc

Node RPC builder

RpcBuilder

Builder for Rpc.

StorageMonitor

Storage monitor

TargetConnections

Wrapper with default value for number of target connections

Enums

BaseBuilderError

Error type for BaseBuilder

BlocksPruning

Block pruning settings.

DsnBuilderError

Error type for DsnBuilder

Event

ExecutionStrategy

Strategy for executing a call into the runtime.

NetworkBuilderError

Error type for NetworkBuilder

OffchainWorkerBuilderError

Error type for OffchainWorkerBuilder

PruningMode

Pruning mode.

Role

Role of the local node.

RpcBuilderError

Error type for RpcBuilder

RpcMethods

Available RPC methods.

Functions

start_announcements_processor

Start processing announcements received by the network node, returns handle that will stop processing on drop.

Type Definitions

ChainSpec

Chain spec type for the subspace

FarmerProviderStorage

Farmer provider storage

NodePieceCache

Node piece cache

NodeProviderStorage

Node provider storage

PieceStore

Farmer piece store

ProviderStorage

General provider storage

RewardsEvent

Events from subspace pallet

SubspaceEvent

Events from subspace pallet