Module subspace_sdk::node

Node related things located here

Re-exports

pub use super::Node;

Modules

chain_spec

Subspace chain configurations.

Structs

BaseBuilder

Builder for Base.

Constraints

Pruning constraints. If none are specified pruning is

DomainConfigBuilder

Builder for DomainConfig.

Dsn

Node DSN builder

DsnBuilder

Builder for Dsn.

DsnOptions

Options for DSN

DsnShared

Shared Dsn structure between node and farmer

ImplName

Type wrapper with default value for implementation name

ImplVersion

Type wrapper with default value for implementation version

InConnections

Wrapper with default value for number of incoming connections

ListenAddresses

Wrapper with default value for listen address

MaxSubsPerConn

Wrapper with default value for max subscriptions per connection

Network

Node network builder

NetworkBuilder

Builder for Network.

OffchainWorker

Offchain worker config

OffchainWorkerBuilder

Builder for OffchainWorker.

OutConnections

Wrapper with default value for number of outgoing connections

PendingInConnections

Wrapper with default value for number of pending incoming connections

PendingOutConnections

Wrapper with default value for number of pending outgoing connections

Rpc

Node RPC builder

RpcBuilder

Builder for Rpc.

StorageMonitor

Storage monitor

SyncingProgress

Current syncing progress

TargetConnections

Wrapper with default value for number of target connections

Enums

BaseBuilderError

Error type for BaseBuilder

BlocksPruning

Block pruning settings.

DsnBuilderError

Error type for DsnBuilder

Event

NetworkBuilderError

Error type for NetworkBuilder

OffchainWorkerBuilderError

Error type for OffchainWorkerBuilder

PruningMode

Pruning mode.

Role

Role of the local node.

RpcBuilderError

Error type for RpcBuilder

RpcMethods

Available RPC methods.

Type Aliases

BlockNumber

Block number

ChainSpec

Chain spec type for the subspace

Hash

Hash type

LocalRecordProvider

A record provider that uses farmer piece cache underneath

RewardsEvent

Events from subspace pallet

SubspaceEvent

Events from subspace pallet